Rename to authentication-popups and generalize
The name `feathers-authentication-popups` isn’t really necessary, so it’s now renamed to `authentication-popups`.

The `middleware.js` file has been renamed to `express.js` to make room for other server frameworks.

The express middleware no longer accepts `app` as an argument, but a cookie config object or a string with the cookie name.